Fix: test errors run by root Fix: package.xml Change: API functions now return NULL on arg parsing error(common convention)
git-svn-id: https://svn.php.net/repository/pecl/eio/trunk@320059 c90b9560-bf6c-de11-be94-00142212c4b1

+++ tests/eio_mkdir_basic.phpt
@@ -3,24 +3,24 @@ Check for eio_mkdir function basic behaviour
 --FILE--
 <?php 
 $temp_dirname = dirname(__FILE__) ."/eio-temp-dir";
-@rmdir($temp_dirname);
+if (file_exists($temp_dirname)) {
+	rmdir($temp_dirname);
+}
 
 function my_mkdir_callback($data, $result) {
 	global $temp_dirname;
 	if ($result == 0 && is_dir($temp_dirname) 
-		&& !is_readable($temp_dirname)
-		&& is_writable($temp_dirname)) {
-		echo "eio_mkdir_ok";
-	}
+		&& fileperms($temp_dirname) & 0300) {
+			echo "eio_mkdir_ok";
+		} else {
+			echo "eio_mkdir_failed";
+		}
 	if (file_exists($temp_dirname))
 		rmdir($temp_dirname);
 }
 
 eio_mkdir($temp_dirname, 0300, 1, "my_mkdir_callback", NULL);
-//eio_event_loop();
-?>
---CLEAN--
-<?php
+eio_event_loop();
 ?>
 --EXPECT--
 eio_mkdir_ok
